+++ src/state.js
@@ -149,7 +149,9 @@ function newGameState() {
             starPowerShoots: 0,
             portraitPath: getPerformerPortraitPath(performer),
             fatigue: 0,
-            loyalty: CONFIG.performers.starting_loyalty
+            loyalty: CONFIG.performers.starting_loyalty,
+            lastBookedDay: null,
+            lastLoyaltyDecayDay: null
           };
         }),
     },